ad_analyzer/requirements.txt
streamlit
openai
transformers
torch
Pillow
python-dotenv